media-source: rename filename using kebab
